+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## unreleased
 ### Added
 * Added `material.OperationXYZ` for merged XYZ operations.
-* Added `material.AssignmentValueXyz.parameter()` method and associated `material.AssignmentValueParameterXyz` type.
+* Added `material.AssignmentValueXyz.parameter` method and associated `material.AssignmentValueParameterXyz` type.
+* Added `material.MaterialParameters.get_parameter` method for converting parameters to floats.
 
 ### Changed
 * Renamed `shader_database.Dependency` to `shader_database.Value` and updated associated type names.